Added a list of 'TimelineEvent' to the Replay object, which stores a TimeSpan of when the event occurred, an enum of the event type, the Player ID of whatever player is involved, and a decimal 'value' if the event has data associated.
I'm intending to use this for various interesting events in the tracker events and game events.  Currently the only one I'm gathering is when the dragon is acquired on Dragon Shire, and I haven't verified it :) I'll do more verification when I'm ready to show this information on HOTSLogs.com.

Also cleaned up the beginning of replay.initData, though there doesn't seem to be any interesting information in these fields.