Ready for patch 34190
Accommodates an existing bug in replay timestamps for the Mac client of Heroes of the Storm.  The timestamps seem to be random (1970, 1999, etc).  I just put the timestamp in the middle of the week for these patches.  I expect Blizzard will fix this soon, as these strange times are also displayed in the client for these replays.

Correctly set data type to 'blob' for game events with this data, so you can easily do a 'ToString()' on these event structures and see some data